+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added a sessionStorage remove item when a new impersonation or stop impersonation is called
+
 ## [1.4.0] - 2022-04-08
 
 ### Added